+++ src/kem/kyber/PERFORMANCE.md
