+++ src/kem/kyber/parameters.rs
@@ -1,3 +1,5 @@
+use crate::digest::{digest_size, Algorithm};
+
 /// Field modulus: 3329
 pub(crate) const FIELD_MODULUS: i32 = 3329;
 
@@ -19,9 +21,20 @@ pub(crate) const BYTES_PER_RING_ELEMENT: usize = BITS_PER_RING_ELEMENT / 8;
 /// this choice.
 pub(crate) const REJECTION_SAMPLING_SEED_SIZE: usize = 168 * 5;
 
-pub(crate) const CPA_PKE_MESSAGE_SIZE: usize = 32;
+/// PKE message size
+pub(crate) const SHARED_SECRET_SIZE: usize = 32;
 
 pub(crate) const CPA_PKE_KEY_GENERATION_SEED_SIZE: usize = 32;
+
+/// Compute serialized length for output size of ByteEncode
+const fn serialized_len<const K: usize, const OUT_LEN: usize>() -> usize {
+    OUT_LEN * K
+}
+
+/// Compute block length for output block size of ByteEncode u (c1)
+const fn block_len<const FACTOR: usize>() -> usize {
+    (COEFFICIENTS_IN_RING_ELEMENT * FACTOR) / 8
+}
 
 /// Rank
 pub(crate) const RANK_512: usize = 2;
@@ -44,16 +57,6 @@ pub(crate) const T_AS_NTT_ENCODED_SIZE_1024: usize =
 pub(crate) const VECTOR_U_COMPRESSION_FACTOR_512: usize = 10;
 pub(crate) const VECTOR_U_COMPRESSION_FACTOR_768: usize = 10;
 pub(crate) const VECTOR_U_COMPRESSION_FACTOR_1024: usize = 11;
-
-/// Compute serialized length for output size of ByteEncode
-const fn serialized_len<const K: usize, const OUT_LEN: usize>() -> usize {
-    OUT_LEN * K
-}
-
-/// Compute block length for output block size of ByteEncode u (c1)
-const fn block_len<const FACTOR: usize>() -> usize {
-    (COEFFICIENTS_IN_RING_ELEMENT * FACTOR) / 8
-}
 
 // Resulting c1 block lengths
 pub(crate) const C1_BLOCK_SIZE_512: usize = block_len::<VECTOR_U_COMPRESSION_FACTOR_512>();
@@ -108,15 +111,13 @@ pub(crate) const CPA_PKE_CIPHERTEXT_SIZE_768: usize =
 pub(crate) const CPA_PKE_CIPHERTEXT_SIZE_1024: usize =
     VECTOR_U_ENCODED_SIZE_1024 + VECTOR_V_ENCODED_SIZE_1024;
 
-pub(crate) const CPA_SERIALIZED_KEY_LEN_512: usize = CPA_PKE_SECRET_KEY_SIZE_768
-    + CPA_PKE_PUBLIC_KEY_SIZE_768
-    + super::hash_functions::H_DIGEST_SIZE
-    + CPA_PKE_MESSAGE_SIZE;
-pub(crate) const CPA_SERIALIZED_KEY_LEN_768: usize = CPA_PKE_SECRET_KEY_SIZE_768
-    + CPA_PKE_PUBLIC_KEY_SIZE_768
-    + super::hash_functions::H_DIGEST_SIZE
-    + CPA_PKE_MESSAGE_SIZE;
-pub(crate) const CPA_SERIALIZED_KEY_LEN_1024: usize = CPA_PKE_SECRET_KEY_SIZE_768
+pub(crate) const H_DIGEST_SIZE: usize = digest_size(Algorithm::Sha3_256);
+
+pub(crate) const SECRET_KEY_SIZE_512: usize =
+    CPA_PKE_SECRET_KEY_SIZE_512 + CPA_PKE_PUBLIC_KEY_SIZE_512 + H_DIGEST_SIZE + SHARED_SECRET_SIZE;
+pub(crate) const SECRET_KEY_SIZE_768: usize =
+    CPA_PKE_SECRET_KEY_SIZE_768 + CPA_PKE_PUBLIC_KEY_SIZE_768 + H_DIGEST_SIZE + SHARED_SECRET_SIZE;
+pub(crate) const SECRET_KEY_SIZE_1024: usize = CPA_PKE_SECRET_KEY_SIZE_1024
     + CPA_PKE_PUBLIC_KEY_SIZE_1024
-    + super::hash_functions::H_DIGEST_SIZE
-    + CPA_PKE_MESSAGE_SIZE;
+    + H_DIGEST_SIZE
+    + SHARED_SECRET_SIZE;
